% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_etwfe.R
\name{adapter_etwfe_poisson}
\alias{adapter_etwfe_poisson}
\title{ETWFE Poisson Adapter (Convenience Wrapper)}
\usage{
adapter_etwfe_poisson()
}
\value{
An estimator_adapter for ETWFE Poisson estimation
}
\description{
A convenience adapter that defaults to Poisson family.
Equivalent to using adapter_etwfe() with family = "poisson".
Note: Requires pop_var and outcome_type to be specified for proper scaling.
}
