% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_etwfe_poisson_glm.R
\name{adapter_etwfe_poisson_glm}
\alias{adapter_etwfe_poisson_glm}
\title{ETWFE Poisson Adapter (Saturated Wooldridge 2023)}
\usage{
adapter_etwfe_poisson_glm()
}
\value{
An estimator_adapter registered as "etwfe_poisson"
}
\description{
Uses fixest::fepois() with saturated cohort x time treatment indicators
and observation-proportional aggregation weights for the overall ATT.
}
