% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_base.R
\name{check_adapter_deps}
\alias{check_adapter_deps}
\title{Check Package Dependencies}
\usage{
check_adapter_deps(adapter)
}
\arguments{
\item{adapter}{An estimator_adapter object}
}
\value{
TRUE if dependencies available, throws error otherwise
}
\description{
Check Package Dependencies
}
