% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_base.R
\name{clear_adapter_registry}
\alias{clear_adapter_registry}
\title{Clear Adapter Registry (For Testing)}
\usage{
clear_adapter_registry()
}
\description{
Clear Adapter Registry (For Testing)
}
\keyword{internal}
