% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{create_wide}
\alias{create_wide}
\title{Create Wide Format Dataset}
\usage{
create_wide(df, group, unit, time, outcome)
}
\arguments{
\item{df}{Long format dataset}

\item{group}{Group variable}

\item{unit}{Unit identifier}

\item{time}{Time variable}

\item{outcome}{Outcome variable}
}
\description{
Create Wide Format Dataset
}
