% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_base.R
\docType{data}
\name{.adapter_registry}
\alias{.adapter_registry}
\title{Adapter Registry}
\format{
An object of class \code{environment} of length 5.
}
\usage{
.adapter_registry
}
\description{
Global registry for storing estimator adapters
}
\keyword{datasets}
