% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_base.R
\name{get_adapter}
\alias{get_adapter}
\title{Get Registered Adapter}
\usage{
get_adapter(name)
}
\arguments{
\item{name}{Character. Adapter name}
}
\value{
An estimator_adapter object
}
\description{
Get Registered Adapter
}
