% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load Power Analysis Configuration}
\usage{
load_config(config_path)
}
\arguments{
\item{config_path}{Path to YAML config file}
}
\description{
Load Power Analysis Configuration
}
