% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_base.R
\name{register_adapter}
\alias{register_adapter}
\title{Register an Adapter}
\usage{
register_adapter(adapter)
}
\arguments{
\item{adapter}{An estimator_adapter object}
}
\description{
Register an Adapter
}
