% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{sample_lognormal_with_mean}
\alias{sample_lognormal_with_mean}
\title{Sample from Lognormal Distribution with Target Mean}
\usage{
sample_lognormal_with_mean(n, desired_mean, cv = 0.5)
}
\arguments{
\item{n}{Number of samples}

\item{desired_mean}{Target mean}

\item{cv}{Coefficient of variation}
}
\description{
Sample from Lognormal Distribution with Target Mean
}
