% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_cs.R
\name{adapter_cs}
\alias{adapter_cs}
\title{Callaway-Sant'Anna Adapter}
\usage{
adapter_cs()
}
\value{
An estimator_adapter for Callaway-Sant'Anna estimation
}
\description{
Translates between staggeredpower's common interface and the \code{did} package
}
\examples{
\donttest{
adapter <- adapter_cs()
result <- adapter$fit(data = my_data, outcome_var = "y", ...)
std_result <- adapter$extract(result)
}
}
