% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_te.R
\name{compute_te}
\alias{compute_te}
\title{Compute Treatment Effects}
\usage{
compute_te(df, period, cohort, group, time, max_year = NULL, outcome)
}
\arguments{
\item{df}{Dataset}

\item{period}{Pre-period type}

\item{cohort}{Control group type}

\item{group}{Group variable}

\item{time}{Time variable}

\item{max_year}{Maximum year}

\item{outcome}{Outcome variable}
}
\description{
Compute Treatment Effects
}
