% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_te.R
\name{compute_te}
\alias{compute_te}
\title{Compute Treatment Effects}
\usage{
compute_te(
  df,
  pta_type,
  enforce_type,
  group_var,
  time_var,
  rel_pass_var,
  outcome
)
}
\arguments{
\item{df}{Dataset}

\item{pta_type}{Which pta assumption ("cs" or "imputation")}

\item{enforce_type}{Method for enforcing parallel trends}

\item{time_var}{time_var variable}

\item{rel_pass_var}{rel_pass_var variable}

\item{outcome}{Outcome variable}

\item{group}{Group variable}
}
\description{
Compute Treatment Effects
}
