% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_models.R
\name{estimate_models}
\alias{estimate_models}
\title{Estimate Difference-in-Differences Models}
\usage{
estimate_models(
  data,
  id_var,
  outcome_var,
  time_var,
  group_var,
  controls,
  models_to_run,
  cluster_var = "state",
  treat_ind_var = "law_pass",
  event_study = FALSE
)
}
\arguments{
\item{data}{Panel dataset}

\item{id_var}{Unit identifier variable}

\item{outcome_var}{Outcome variable}

\item{controls}{Whether to include controls}

\item{models_to_run}{Vector of models to estimate}

\item{event_study}{Whether to run event study}
}
\description{
Estimate Difference-in-Differences Models
}
