% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_models.R
\name{estimate_models}
\alias{estimate_models}
\title{Estimate Difference-in-Differences Models}
\usage{
estimate_models(
  data,
  id_var,
  outcome_var,
  arrest_law_type,
  use_controls,
  models_to_run,
  event_study = FALSE
)
}
\arguments{
\item{data}{Panel dataset}

\item{id_var}{Unit identifier variable}

\item{outcome_var}{Outcome variable}

\item{arrest_law_type}{Type of law}

\item{use_controls}{Whether to include controls}

\item{models_to_run}{Vector of models to estimate}

\item{event_study}{Whether to run event study}
}
\description{
Estimate Difference-in-Differences Models
}
