% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{run_power_analysis}
\alias{run_power_analysis}
\title{Run Power Analysis with Parallel Trends}
\usage{
run_power_analysis(
  data_clean,
  unit_var,
  group_var,
  time_var,
  rel_pass_var,
  treat_ind_var,
  controls = NULL,
  outcome,
  transform_outcome = NULL,
  pta_type,
  enforce_type = NULL,
  percent_effect,
  models_to_run = c("cs", "imputation", "twfe"),
  n_sims = 100
)
}
\arguments{
\item{data_clean}{Clean panel dataset}

\item{unit_var}{Level of analysis (state or county)}

\item{controls}{list of controls}

\item{outcome}{Outcome variable}

\item{pta_type}{Which pta assumption ("cs" or "imputation")}

\item{enforce_type}{Method for enforcing parallel trends}

\item{percent_effect}{Effect size to simulate}

\item{n_sims}{Number of simulations}
}
\description{
Run Power Analysis with Parallel Trends
}
