% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_results.R
\name{store_results}
\alias{store_results}
\title{Store Power Analysis Results}
\usage{
store_results(
  results,
  data,
  aggregate_results,
  event_study_results,
  pta_type,
  enforce_type,
  analysis_level,
  outcome,
  use_controls,
  drop_add_states,
  result_type,
  transform_outcome
)
}
\arguments{
\item{results}{Analysis results}

\item{data}{Original dataset}

\item{aggregate_results}{Aggregated results}

\item{event_study_results}{Event study results}

\item{pta_type}{Pta type}

\item{enforce_type}{Enforce type}

\item{analysis_level}{Analysis level}

\item{outcome}{Outcome variable}

\item{use_controls}{Control usage}

\item{drop_add_states}{State dropping}

\item{result_type}{Result type}

\item{model}{Model type}
}
\description{
Store Power Analysis Results
}
